greedy v3.0 it works

+++ solver.h
@@ -6,7 +6,7 @@
 #define OUTPUT_DIRECTORY "./solutions/"
 //typedef std::map<int,int> hash;
 typedef std::pair<char, int> nodeKey;
-typedef std::pair<std::pair<int, int>, std::pair<double, double>> AFSDepotRouteInfo;
+typedef std::pair<std::pair<int, double>, std::pair<double, double>> AFSDepotRouteInfo;
 
 double distanceHarvesine(double lon1, double lat1, double lon2, double lat2);
 void printNodeKeyVector(std::vector<nodeKey> v);
